+++ moabb/datasets/schirrmeister2017.py
@@ -17,15 +17,6 @@ GIN_URL = (
 class Schirrmeister2017(BaseDataset):
     """High-gamma dataset described in Schirrmeister et al. 2017.
 
-    .. admonition:: Dataset summary
-
-
-        =================  =======  =======  ==========  =================  ============  ===============  ===========
-        Name                 #Subj    #Chan    #Classes    #Trials / class  Trials len    Sampling rate      #Sessions
-        =================  =======  =======  ==========  =================  ============  ===============  ===========
-        Schirrmeister2017       14      128           4                120  4s            500Hz                      1
-        =================  =======  =======  ==========  =================  ============  ===============  ===========
-
     Dataset from [1]_
 
     Our “High-Gamma Dataset” is a 128-electrode dataset (of which we later only use